# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proshli API",
    "description": "Russian-language job aggregator — backend API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/resumes/upload": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Upload Resume",
        "operationId": "resumes-upload_resume",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_resumes-upload_resume"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resumes/versions": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Create Resume Version",
        "operationId": "resumes-create_resume_version",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeVersionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeVersionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "resumes"
        ],
        "summary": "List Resume Versions",
        "operationId": "resumes-list_resume_versions",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeVersionOut"
                  },
                  "title": "Response Resumes-List Resume Versions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resumes/versions/{version_id}/improve": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Improve Resume Version",
        "description": "Return an LLM-tightened summary + bullet suggestions for one version.\n\nSeeker-only; counts against the same per-day AI budget as ``/ai/chat``\nso the pro/free tier caps stay honest. The version blob is read from\n``content_json`` and passed to the LLM as-is — the model decides what's\nworth rewriting.",
        "operationId": "resumes-improve_resume_version",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Id"
            }
          },
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeImproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeImproveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/digest/preferences": {
      "put": {
        "tags": [
          "digest"
        ],
        "summary": "Update Digest Preferences",
        "operationId": "digest-update_digest_preferences",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DigestPreferenceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestPreferenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "digest"
        ],
        "summary": "Get Digest Preferences",
        "operationId": "digest-get_digest_preferences",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestPreferenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "digest"
        ],
        "summary": "Disable Digest",
        "description": "Turn off the digest entirely for the current user.\n\nBot calls this from ``/digest_off`` so the seeker can opt out without\nknowing the channel/frequency knobs. We keep the row (so we don't\nlose ``telegram_chat_id`` for if they re-enable later) and just flip\nboth transports off — this is semantically what the worker checks\nbefore sending.",
        "operationId": "digest-disable_digest",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/digest/preview": {
      "get": {
        "tags": [
          "digest"
        ],
        "summary": "Preview Digest",
        "operationId": "digest-preview_digest",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proshli_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Proshli Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestPreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_resumes-upload_resume": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_resumes-upload_resume"
      },
...
      "DigestItem": {
        "properties": {
          "vacancy_id": {
            "type": "integer",
            "title": "Vacancy Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "score_reason": {
            "type": "string",
            "title": "Score Reason"
          }
        },
        "type": "object",
        "required": [
          "vacancy_id",
          "title",
          "company",
          "location",
          "score_reason"
        ],
        "title": "DigestItem"
      },
      "DigestPreferenceOut": {
        "properties": {
          "frequency": {
            "type": "string",
            "enum": [
              "daily",
              "weekly"
            ],
            "title": "Frequency"
          },
          "via_telegram": {
            "type": "boolean",
            "title": "Via Telegram"
          },
          "via_email": {
            "type": "boolean",
            "title": "Via Email"
          },
          "telegram_chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Chat Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "frequency",
          "via_telegram",
          "via_email",
          "telegram_chat_id",
          "updated_at"
        ],
        "title": "DigestPreferenceOut"
      },
      "DigestPreferenceUpdate": {
        "properties": {
          "frequency": {
            "type": "string",
            "enum": [
              "daily",
              "weekly"
            ],
            "title": "Frequency",
            "default": "daily"
          },
          "via_telegram": {
            "type": "boolean",
            "title": "Via Telegram",
            "default": true
          },
          "via_email": {
            "type": "boolean",
            "title": "Via Email",
            "default": false
          },
          "telegram_chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Chat Id"
          }
        },
        "type": "object",
        "title": "DigestPreferenceUpdate"
      },
      "DigestPreviewOut": {
        "properties": {
          "frequency": {
            "type": "string",
            "enum": [
              "daily",
              "weekly"
            ],
            "title": "Frequency"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Channels"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DigestItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "frequency",
          "channels",
          "items"
        ],
        "title": "DigestPreviewOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeImproveRequest": {
        "properties": {
          "target_role": {
            "type": "string",
            "maxLength": 128,
            "title": "Target Role",
            "default": ""
          },
          "focus": {
            "type": "string",
            "maxLength": 500,
            "title": "Focus",
            "default": ""
          }
        },
        "type": "object",
        "title": "ResumeImproveRequest",
        "description": "Optional hint payload for ``POST /resumes/versions/{id}/improve``.\n\nBoth fields are optional — the endpoint defaults to the version's own\n``target_role`` if ``target_role`` is empty, and skips the per-turn\nfocus block if ``focus`` is empty."
      },
      "ResumeImproveResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "used_today": {
            "type": "integer",
            "title": "Used Today"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "suggestions",
          "used_today",
          "limit",
          "backend"
        ],
        "title": "ResumeImproveResponse",
        "description": "LLM-improved resume summary + concrete suggestions.\n\n``summary`` is a 1–2 sentence pitch the seeker can paste at the top of\na resume / cover letter. ``suggestions`` is a short bulleted list of\nactionable rewrites (each item ≤ 1 sentence). ``used_today`` / ``limit``\nmirror the AI-budget surface so the bot can warn the user when they're\nclose to the cap. ``backend`` is ``\"anthropic\"`` or ``\"rule_based\"`` so\nthe FE / bot can distinguish a real LLM call from the offline fallback."
      },
      "ResumeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parsed_skills": {
            "type": "string",
            "title": "Parsed Skills"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "parsed_skills",
          "created_at"
        ],
        "title": "ResumeOut"
      },
      "ResumeVersionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "target_role": {
            "type": "string",
            "title": "Target Role",
            "default": ""
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "name",
          "content"
        ],
        "title": "ResumeVersionCreate"
      },
      "ResumeVersionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_role": {
            "type": "string",
            "title": "Target Role"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "target_role",
          "content",
          "created_at"
        ],
        "title": "ResumeVersionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}